Sophie HW files & tokenization
Some errors I encounter are punctuation issues. Tokenization includes punctuation, which splits the words. For example "Achilles!" is split into "Achilles" and "!" so some words may be hyphenated or contain apostrophes affecting tokenization. Some words may also be treated as separate tokens if they differ in the use of a capital letter. Line breaks and footnotes or spaces between stanzas may also interfere with tokenization. In order to refine the tokenization process, all text should be converted to lowercase to avoid case differences and punctuation should all be removed.